backend/app/pyproject.toml
[tool.poetry]
name = "avault"
version = "0.1.0"
description = "Instant Messaging Website"
authors = ["Agnirudra Sil <agnirudra.sil@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.11.3"
fastapi = "^0.54.1"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.27.1"
celery = "^4.4.2"
passlib = { extras = ["bcrypt"], version = "^1.7.2" }
tenacity = "^6.1.0"
pydantic = "^1.4"
emails = "^0.5.15"
raven = "^6.10.0"
gunicorn = "^20.0.4"
jinja2 = "^2.11.2"
psycopg2-binary = "^2.8.5"
alembic = "^1.4.2"
sqlalchemy = "^1.3.16"
pytest = "^5.4.1"
python-jose = { extras = ["cryptography"], version = "^3.1.0" }
aiohttp = "^3.8.1"
aioredis = "^2.0.1"
aio-pika = "^6.8.1"
msgpack = "^1.0.3"
argon2_cffi = "^21.3.0"
python-dotenv = "^0.19.2"
emoji = "^1.7.0"
beautifulsoup4 = "^4.11.1"
mistune = "^2.0.2"
asgiref = "^3.5.0"
redis = "^4.2.2"
boto3 = "^1.21.40"
Werkzeug = "^2.1.1"
Pillow = "^9.1.0"
limits = "^2.6.1"
coredis = "^3.4.6"
pyotp = "^2.6.0"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
